+++ R/tree_distance_utilities.R
@@ -9,6 +9,7 @@
 #' @template MRS
 #' @keywords internal
 #' @importFrom TreeTools as.Splits TipLabels
+#' @importFrom utils combn
 #' @export
 CalculateTreeDistance <- function (Func, tree1, tree2 = NULL,
                                    reportMatching = FALSE, ...) {
@@ -80,18 +81,13 @@ CalculateTreeDistance <- function (Func, tree1, tree2 = NULL,
                                     nTip = length(tipLabels), ...) {
   splits <- as.Splits(splits1, tipLabels = tipLabels, asSplits = FALSE)
   nSplits <- length(splits)
-  notLastSplit <- nSplits - 1L
-  ret <- matrix(0, nSplits, nSplits)
-  is <- matrix(seq_len(nSplits), nSplits, nSplits)
+  is <- combn(seq_len(nSplits), 2)
   
   ret <- structure(class = 'dist', Size = nSplits,
                    diag = FALSE, upper = FALSE,
-                   mapply(Func,
-                          splits[t(is)[lower.tri(is)]],
-                          splits[is[lower.tri(is)]],
-                          nTip = nTip,
-                          reportMatching = FALSE,
-                          ...))
+                   apply(is, 2, function (i)
+                     Func(splits[[i[1]]], splits[[i[2]]],
+                          nTip = nTip, reportMatching = FALSE, ...)))
   # Return:
   ret
 }
@@ -219,7 +215,8 @@ CalculateTreeDistance <- function (Func, tree1, tree2 = NULL,
 #' @export
 Entropy <- function (...) {
   p <- c(...)
-  -sum(p[p > 0] * log2(p[p > 0]))
+  p <- p[p > 0]
+  -sum(p * log2(p))
 }
 
 #' Distances between each pair of trees